third_party/WebKit/LayoutTests/fast/forms/input-stepup-stepdown-expected.txt

Check stepUp() and stepDown() bahevior for unsupported types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Unsupported type
PASS input.step = "3"; input.min = ""; input.max = ""; input.value = "2"; input.stepDown() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS input.stepDown(0) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS input.stepUp() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS input.stepUp(0) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS successfullyParsed is true

TEST COMPLETE
